api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
pydantic==2.5.3
httpx==0.26.0
PyJWT==2.8.0
bcrypt==4.1.2
confluent-kafka==2.3.0
pytest==7.4.4
pytest-mock==3.12.0
